Moving terminals and icons related files to FMI/*
