+++ src/Import/include/FMI3/fmi3_import.h
@@ -35,7 +35,7 @@
 #include "fmi3_import_unit.h"
 #include "fmi3_import_variable.h"
 #include "fmi3_import_variable_list.h"
-#include "fmi3_import_terminals_and_icons.h"
+#include <FMI/fmi_import_terminals_and_icons.h>
 
 #include "fmi3_import_capi.h"
 #include "fmi3_import_convenience.h"
