+++ src/Import/src/FMI2/fmi2_import_impl.h
@@ -24,7 +24,7 @@
 
 #include <FMI2/fmi2_import.h>
 #include <FMI2/fmi2_xml_model_description.h>
-#include "FMI3/fmi3_xml_terminals_and_icons.h"
+#include <FMI/fmi_xml_terminals_and_icons.h>
 
 #include "../FMI/fmi_import_context_impl.h"
 #include "../src/FMI2/fmi2_capi_impl.h"
