+++ src/XML/src/FMI3/fmi3_xml_parser_context_impl.h
@@ -20,9 +20,8 @@
 #include <JM/jm_vector.h>
 #include <JM/jm_stack.h>
 
-// #include <FMI3/fmi3_enums.h>
 #include <FMI3/fmi3_xml_model_description.h>
-#include <FMI3/fmi3_xml_terminals_and_icons.h>
+#include <FMI/fmi_xml_terminals_and_icons.h>
 #include "fmi3_xml_parser_lists.h"
 
 #ifdef __cplusplus
